rename to `webkitgtk` on linux
